fix(auth): refactor auth flow and fix jotai provider issue
- Move Jotai Provider to root layout to fix multiple instances warning
- Handle redirect URL logic inside AuthClient component
- Remove duplicate redirect logic from Login and Register pages
- Update AuthClient to use useRouter and redirectAtom state
- Clean up unused imports and commented out code
